feat: add subscribe and contact route
