--- src/components/ContactComponent.jsx
@@ -1,130 +1,0 @@
-import React from "react";
-import { FaHome, FaAngleRight, FaSearch } from "react-icons/fa";
-
-const ContactComponent = () => {
-  return (
-    <section className="bg-white pt-10 p-6 rounded-lg shadow-md mx-4 sm:mx-10 md:mx-20 lg:mx-10">
-      <nav className="flex mb-3" aria-label="Breadcrumb">
-        <ol className="inline-flex items-center space-x-1 md:space-x-2 rtl:space-x-reverse">
-          <li className="inline-flex items-center">
-            <a
-              href="/"
-              className="inline-flex items-center text-sm font-medium text-gray-700 hover:text-blue-600 dark:text-gray-400 dark:hover:text-white"
-            >
-              <FaHome className="mr-2" />
-              Home
-            </a>
-          </li>
-
-          <li aria-current="page">
-            <div className="flex items-center">
-              <FaAngleRight />
-              <span className="ms-1 text-sm font-medium text-gray-500 md:ms-2 dark:text-gray-400">
-                Contact
-              </span>
-            </div>
-          </li>
-        </ol>
-      </nav>
-      <div className="flex items-center flex-column flex-wrap md:flex-row justify-between mb-4">
-        <h5 className="text-xl font-bold text-gray-600 dark:text-white">
-          Contact
-        </h5>
-        <div className="flex items-center space-x-4">
-          <div className="relative">
-            <div className="absolute inset-y-0 left-0 rtl:inset-r-0 rtl:right-0 flex items-center ps-3 pointer-events-none">
-              <FaSearch />
-            </div>
-            <input
-              type="text"
-              id="table-search"
-              className="block p-2 ps-10 text-sm text-gray-900 border border-gray-300 rounded-lg w-50 h-8 bg-gray-50 focus:ring-blue-500 focus:border-blue-500 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
-              placeholder="Search"
-            />
-          </div>
-        </div>
-      </div>
-      <div className="overflow-x-auto">
-        <table className="w-full text-sm text-left rtl:text-right text-gray-500 dark:text-gray-400 overflow-x-auto">
-          <thead className="text-xs text-gray-700 uppercase bg-gray-50 dark:bg-gray-700 dark:text-gray-400">
-            <tr>
-              <th className="border-b py-3 px-6">No</th>
-              <th className="border-b py-3 px-6">Name</th>
-              <th className="border-b py-3 px-6">Email</th>
-              <th className="border-b py-3 px-6">Phone</th>
-              <th className="border-b py-3 px-6">Message</th>
-              <th className="border-b py-3 px-6">Date</th>
-            </tr>
-          </thead>
-          <tbody>
-            <tr className="bg-white border-b dark:bg-gray-800 dark:border-gray-700 hover:bg-gray-50 dark:hover:bg-gray-600">
-              <td className="border-b py-4 px-6">1</td>
-              <td className="border-b py-4 px-6">Deni Irawan Nugraha</td>
-              <td className="border-b py-4 px-6">deni.irawan@lumoshive.co</td>
-              <td className="border-b py-4 px-6">083174869601</td>
-              <td className="border-b py-4 px-6">
-                consectetur adipisicing elit. Reprehenderit, architecto ab
-                molestias, quasi, fuga fugiat ipsa consectetur doloremque
-                inventore veritatis explicabo at error maiores tempore rem totam
-                animi provident in.
-              </td>
-              <td className="border-b py-2">2024-11-16T02:16:44.240</td>
-            </tr>
-          </tbody>
-        </table>
-      </div>
-      <nav
-        className="flex items-center flex-column flex-wrap md:flex-row justify-between pt-4"
-        aria-label="Table navigation"
-      >
-        <span className="text-sm font-normal text-gray-500 dark:text-gray-400 mb-4 md:mb-0 block w-full md:inline md:w-auto">
-          Showing{" "}
-          <span className="font-semibold text-gray-900 dark:text-white">
-            1-10
-          </span>{" "}
-          of{" "}
-          <span className="font-semibold text-gray-900 dark:text-white">
-            1000
-          </span>
-        </span>
-        <ul className="inline-flex -space-x-px rtl:space-x-reverse text-sm h-8">
-          <li>
-            <a
-              href="#"
-              className="flex items-center justify-center px-2 h-8 ms-0 leading-tight text-gray-500 bg-white border border-gray-300 rounded-s-lg hover:bg-gray-100 hover:text-gray-700 dark:bg-gray-800 dark:border-gray-700 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white"
-            >
-              Previous
-            </a>
-          </li>
-          <li>
-            <a
-              href="#"
-              className="flex items-center justify-center px-2 h-8 leading-tight text-gray-500 bg-white border border-gray-300 hover:bg-gray-100 hover:text-gray-700 dark:bg-gray-800 dark:border-gray-700 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white"
-            >
-              1
-            </a>
-          </li>
-          <li>
-            <a
-              href="#"
-              className="flex items-center justify-center px-2 h-8 leading-tight text-gray-500 bg-white border border-gray-300 hover:bg-gray-100 hover:text-gray-700 dark:bg-gray-800 dark:border-gray-700 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white"
-            >
-              2
-            </a>
-          </li>
-
-          <li>
-            <a
-              href="#"
-              className="flex items-center justify-center px-2 h-8 leading-tight text-gray-500 bg-white border border-gray-300 rounded-e-lg hover:bg-gray-100 hover:text-gray-700 dark:bg-gray-800 dark:border-gray-700 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white"
-            >
-              Next
-            </a>
-          </li>
-        </ul>
-      </nav>
-    </section>
-  );
-};
-
-export default ContactComponent;
