+++ src/components/Navbar/SidebarComponent.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import { Link, useLocation } from "react-router-dom";
-import { useAuth } from "../context/AuthContext";
+import { useAuth } from "../../context/AuthContext";
 import {
   FaUsers,
   FaBriefcase,
